:heavy_minus_sign: Remove the vendor from repository
The vendor folder is too large for this project. The vendor was included in the project at the beginning because it is convenient for users to download and use directly from the project. It is a very convenient way for people who are inconvenient to use Google services. Under the trade-off, I still think that the vendor is not necessary, and users should deal with network issues by themselves.